third_party/blink/web_tests/fast/dom/HTMLElement/attr-dir-auto-remove-add-children-expected.txt

Test that directionality of an element with dir=auto set is re-evaluated when the first child text node of that element is added or removed. Also test that directionality does not change if the element does not have the attribute dir=auto.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.defaultView.getComputedStyle(parent1, null).getPropertyValue('border-right-color') is 'rgb(0, 128, 0)'
PASS document.defaultView.getComputedStyle(parent2, null).getPropertyValue('border-right-color') is 'rgb(255, 0, 0)'
PASS document.defaultView.getComputedStyle(parent3, null).getPropertyValue('border-right-color') is 'rgb(255, 0, 0)'
PASS document.defaultView.getComputedStyle(parent1, null).getPropertyValue('border-right-color') is 'rgb(255, 0, 0)'
PASS document.defaultView.getComputedStyle(parent2, null).getPropertyValue('border-right-color') is 'rgb(0, 128, 0)'
PASS document.defaultView.getComputedStyle(parent2, null).getPropertyValue('border-right-color') is 'rgb(255, 0, 0)'
PASS document.defaultView.getComputedStyle(parent3, null).getPropertyValue('border-right-color') is 'rgb(255, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE
